requirements.txt
fastf1>=3.6.1
pandas>=1.5
numpy>=1.24
scikit-learn>=1.2
lightgbm>=3.3
xgboost>=1.7
matplotlib>=3.6
seaborn>=0.12
plotly>=5.0
shap>=0.42
streamlit>=1.29
pyyaml>=6.0
